BackEnd/requirements.txt
bidict==0.21.2
click==8.0.1
Flask==2.0.1
Flask-Cors==3.0.10
Flask-HTTPAuth==4.4.0
Flask-SocketIO==5.1.0
gevent==21.1.2
gevent-websocket==0.10.1
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.1
opencv-python==4.5.3.56
paho-mqtt==1.5.1
PyJWT==2.1.0
python-dotenv==0.19.0
python-engineio==4.2.0
python-socketio==5.3.0
six==1.16.0
Werkzeug==2.0.1
zope.event==4.5.0
zope.interface==5.4.0
requests==2.26.0
schedule==1.1.0